Tweak

+++ src/mask.rs
@@ -19,9 +19,9 @@ impl Mask {
       Self::Circle => v.norm() < 1.0,
       Self::Cross => v.x.abs() < 0.25 || v.y.abs() < 0.25,
       Self::Mod { divisor, remainder } => {
-        (pixel.x * state.matrix.nrows() as isize + pixel.y) % *divisor as isize == *remainder as isize
+        (pixel.x as usize * state.matrix.nrows() + pixel.y as usize) % *divisor == *remainder
       }
-      Self::Rows { nrows, step } => pixel.y % (nrows.saturating_add(*step) as isize) < *nrows as isize,
+      Self::Rows { nrows, step } => pixel.y as usize % (nrows.saturating_add(*step)) < *nrows,
       Self::Square => v.abs() < Vector2::new(0.5, 0.5),
       Self::Top => v.y < 0.0,
       Self::X => (v.x - v.y).abs() < 0.25 || (v.x + v.y).abs() < 0.25,
